requirements.txt
Django==4.1
djangorestframework==3.13.1
django-ipware==4.0.2
pathlib==1.0.1
Jinja2==3.1.2
#If you use mongomanager: pymongo
# wkhtml2pdf https://wkhtmltopdf.org/downloads.html
PyPDF2==2.12.1
pdfkit==1.0.0